Module numa_gpu::runtime::cuda

CUDA runtime for data transfer and kernel execution.

There exist multiple methods to transfer data from main-memory to device memory. Also, data transfer and execution should overlap for the best performance. This module provides a collection of transfer method implementations, and efficient iterators for executing GPU kernels.

Structs

CudaIterator2

CUDA iterator for two mutable inputs.

CudaTransferStrategyMeasurement

Timings of the CudaTransferStrategy phases

CudaUnifiedIterator2

CUDA iterator for two mutable unified memory inputs.

EventTimer

Timer based on CUDA events.

Enums

CudaTransferStrategy

Specify the CUDA transfer strategy.

Traits

IntoCudaIterator

Conversion into a CUDA iterator.

IntoCudaIteratorWithStrategy

Conversion into a CUDA iterator with a specified transfer strategy.