Trait numa_gpu::runtime::cuda::IntoCudaIterator

pub trait IntoCudaIterator<'a> {
    type Iter;
    fn into_cuda_iter(&'a mut self, chunk_len: usize) -> Result<Self::Iter>;
}

Conversion into a CUDA iterator.

By implementing IntoCudaIterator for a type, you define how the type is converted into an iterator capable of executing CUDA functions on a GPU.

The iterator must define a transfer strategy.

The chunk_len parameter specifies the granularity of each data transfer from main-memory to device memory. The same granularity is used when passing input parameters to the GPU kernel.

Associated Types

type Iter

The type of the iterator to produce.

Required methods

fn into_cuda_iter(&'a mut self, chunk_len: usize) -> Result<Self::Iter>

Creates an iterator from a value.

See the module-level documentation for details.