Trait numa_gpu::runtime::cuda::IntoCudaIteratorWithStrategy

pub trait IntoCudaIteratorWithStrategy<'a> {
    type Iter;
    fn into_cuda_iter_with_strategy(
        &'a mut self, 
        strategy: CudaTransferStrategy, 
        chunk_len: usize, 
        cpu_memcpy_threads: usize, 
        cpu_affinity: &CpuAffinity
    ) -> Result<Self::Iter>;
}

Conversion into a CUDA iterator with a specified transfer strategy.

By implementing IntoCudaIteratorWithStrategy for a type, you define how the type is converted into an iterator capable of executing CUDA functions on a GPU.

strategy specifies which transfer strategy to use. See the CudaTransferStrategy documation for details. The iterator must implement all strategies.

The chunk_len parameter specifies the granularity of each data transfer from main-memory to device memory. The same granularity is used when passing input parameters to the GPU kernel.

Associated Types

type Iter

The type of the iterator to produce.

Required methods

fn into_cuda_iter_with_strategy(
&'a mut self,
strategy: CudaTransferStrategy,
chunk_len: usize,
cpu_memcpy_threads: usize,
cpu_affinity: &CpuAffinity
) -> Result<Self::Iter>

Creates an iterator from a value.

See the module-level documentation for details.