Module numa_gpu::runtime::cuda_wrapper

Collection of Rust-ified wrappers for commonly-used CUDA functions.

Structs

CudaMemInfo

CUDA memory information

Enums

MemAdviseFlags

Constants

CPU_DEVICE_ID

Functions

async_copy

Copy a slice using CUDA’s memcpyAsync function.

current_device_id

host_register⚠

Page-lock an existing memory range for efficient GPU transfers.

host_unregister⚠

Unregisters a memory range that was page-locked with host_register().

mem_advise

Advise how the memory range will be used.

mem_info

Returns the free and total device memory in bytes

memset_async

Fill a launchable slice using the CUDA memset_async function

prefetch_async

Prefetch memory to the device specified in the current context.