Trait numa_gpu::runtime::hw_info::CudaDeviceInfo

pub trait CudaDeviceInfo {
    fn sm_cores(&self) -> Result<u32>;

    fn cores(&self) -> Result<u32>;

    fn concurrent_managed_access(&self) -> Result<bool>;

    fn clock_rate(&self) -> Result<u32>;

    fn memory_clock_rate(&self) -> Result<u32>;
}

Extends Rustacuda’s Device with methods that provide additional hardware information.

Required methods

fn sm_cores(&self) -> Result<u32>

Returns the number of cores per streaming multiprocessor

fn cores(&self) -> Result<u32>

Returns the total number of cores that are in the device

fn concurrent_managed_access(&self) -> Result<bool>

Returns true if concurrent managed access is supported by the device

fn clock_rate(&self) -> Result<u32>

Returns the default clock rate of the streaming multiprocessor in megahertz

fn memory_clock_rate(&self) -> Result<u32>

Returns the default memory clock rate of the GPU in megahertz