Module numa_gpu::runtime::linux_wrapper

Structs

CpuSet

CPU set to create CPU and NUMA node masks.

MemBindFlags

Flags for mbind

MemPolicyModes

Memory policies

MemProtectFlags

NumaMemInfo

NUMA node memory information

Traits

MemProtect

Defines the mprotect system call for the type

Functions

mbind

numa_mem_info

Returns the memory information of a NUMA node

numa_node_of_cpu

Find the NUMA node that the CPU core belongs to.

numa_run_on_node

Run the current thread on the specified NUMA node.

numa_set_preferred

Try to allocate memory on the sepecified NUMA node.

numa_set_strict

NUMA allocations will fail if the memory cannot be allocated on the target NUMA node. The default behavior is to fall back on other nodes.

numa_tonode_memory

Put memory on a specific node.