Module numa_gpu::runtime::memory

Re-exports

pub use self::Mem::*;

Structs

LaunchableMutPtr

A pointer to mutable memory that can be dereferenced on the GPU.

LaunchableMutSlice

A slice of mutable memory that can be dereferenced on the GPU.

LaunchablePtr

A pointer to immutable memory that can be dereferenced on the GPU.

LaunchableSlice

A slice of immutable memory that can be dereferenced on the GPU.

Enums

DerefMem

A CPU-dereferencable memory type

Mem

MemType

A MemType represents the memory type of Mem

Traits

LaunchableMem

GPU-accessible memory.

MemLock

A trait for locking pages in memory

PageLock

A trait for memory that can be page-locked by CUDA.