Trait numa_gpu::runtime::memory::LaunchableMem

pub trait LaunchableMem {
    type Item;
    fn as_launchable_ptr(&self) -> LaunchablePtr<Self::Item>;

    fn as_launchable_slice(&self) -> LaunchableSlice<'_, Self::Item>;

    fn as_launchable_mut_ptr(&mut self) -> LaunchableMutPtr<Self::Item>;

    fn as_launchable_mut_slice(&mut self) -> LaunchableMutSlice<'_, Self::Item>;
}

GPU-accessible memory.

By implementing LaunchableMem for a type, you specify that the memory can be directly accessed on the GPU.

Associated Types

type Item

The type of elements stored in the memory range.

Required methods

fn as_launchable_ptr(&self) -> LaunchablePtr<Self::Item>

Returns a launchable pointer to the beginning of the memory range.

fn as_launchable_slice(&self) -> LaunchableSlice<'_, Self::Item>

Returns a launchable slice to the entire memory range.

fn as_launchable_mut_ptr(&mut self) -> LaunchableMutPtr<Self::Item>

Returns a launchable mutable pointer to the beginning of the memory range.

fn as_launchable_mut_slice(&mut self) -> LaunchableMutSlice<'_, Self::Item>

Returns a launchable mutable slice to the entire memory range.